requirements.txt
streamlit
torch
transformers
sentence-transformers
PyPDF2
python-docx

